requirements.txt
streamlit
plotly
pandas
python-dotenv
reportlab
schedule
streamlit-cookies-manager
numpy
openai==0.28
